+++ lib/oli/resources/numbering.ex
@@ -37,14 +37,18 @@ defmodule Oli.Resources.Numbering do
   """
   @spec full_hierarchy(resolver, project_or_section_slug) :: [%HierarchyNode{}]
   def full_hierarchy(resolver, project_or_section_slug) do
+
+    revisions_by_id = resolver.all_revisions_in_hierarchy(project_or_section_slug)
+    |> Enum.reduce(%{}, fn r, m -> Map.put(m, r.resource_id, r) end)
+
     full_hierarchy_helper(
       number_full_tree(resolver, project_or_section_slug),
-      resolver.all_revisions_in_hierarchy(project_or_section_slug),
+      revisions_by_id,
       resolver.root_container(project_or_section_slug)
     )
   end
 
-  def full_hierarchy_helper(numberings, revisions, revision) do
+  def full_hierarchy_helper(numberings, revisions_by_id, revision) do
     [
       %HierarchyNode{
         revision: revision,
@@ -54,8 +58,8 @@ defmodule Oli.Resources.Numbering do
             fn resource_id ->
               full_hierarchy_helper(
                 numberings,
-                revisions,
-                Enum.find(revisions, fn rev -> rev.resource_id == resource_id end)
+                revisions_by_id,
+                Map.get(revisions_by_id, resource_id)
               )
             end
           ),