+++ lib/oli_web/controllers/delivery_controller.ex
@@ -1,6 +1,5 @@
 defmodule OliWeb.DeliveryController do
   use OliWeb, :controller
-  import Oli.Utils
 
   alias Oli.Delivery.Sections
   alias Oli.Delivery.Sections.Section
@@ -8,7 +7,7 @@ defmodule OliWeb.DeliveryController do
   alias Oli.Institutions
   alias Lti_1p3.Tool.{PlatformRoles, ContextRoles}
   alias Oli.Accounts
-  alias Oli.Accounts.{Author, User}
+  alias Oli.Accounts.Author
 
   @allow_configure_section_roles [
     PlatformRoles.get_role(:system_administrator),
@@ -317,26 +316,28 @@ defmodule OliWeb.DeliveryController do
     |> redirect(to: Routes.static_page_path(conn, :index))
   end
 
-  def new_user(conn, %{"redirect_to" => redirect_to}) do
-    changeset = Accounts.change_user(%User{})
-    render(conn, "new_user.html", changeset: changeset, redirect_to: redirect_to)
+  def enroll(conn, _params) do
+    section = conn.assigns.section
+    render(conn, "enroll.html", section: section)
   end
 
-  def create_user(conn, %{
-        "user_details" => user_details,
-        "g-recaptcha-response" => g_recaptcha_response
-      }) do
-    redirect_to = value_or(user_details["redirect_to"], Routes.delivery_path(conn, :index))
+  defp recaptcha_verified?(g_recaptcha_response) do
+    g_recaptcha_response != "" and Oli.Utils.Recaptcha.verify(g_recaptcha_response) == {:success, true}
+  end
 
-    with g_recaptcha_response when g_recaptcha_response != "" <- g_recaptcha_response,
-         {:success, true} <- Oli.Utils.Recaptcha.verify(g_recaptcha_response) do
+  def create_user(conn, %{"g-recaptcha-response" => g_recaptcha_response}) do
+
+    if Oli.Utils.LoadTesting.enabled?() or recaptcha_verified?(g_recaptcha_response) do
+
+      section = conn.assigns.section
+
       with {:ok, user} <-
-             Accounts.create_user(%{
-               # generate a unique sub identifier which is also used so a user can access
-               # their progress in the future or using a different browser
-               sub: UUID.uuid4(),
-               guest: true,
-             }) do
+            Accounts.create_user(%{
+              # generate a unique sub identifier which is also used so a user can access
+              # their progress in the future or using a different browser
+              sub: UUID.uuid4(),
+              guest: true,
+            }) do
         Accounts.update_user_platform_roles(user, [
           PlatformRoles.get_role(:institution_learner)
         ])
@@ -344,18 +345,16 @@ defmodule OliWeb.DeliveryController do
         conn
         |> OliWeb.Pow.PowHelpers.use_pow_config(:user)
         |> Pow.Plug.create(user)
-        |> redirect(to: redirect_to)
+        |> redirect(to: Routes.page_delivery_path(conn, :index, section.slug))
       else
-        {:error, changeset} ->
-          render(conn, "new_user.html", changeset: changeset, redirect_to: redirect_to)
+        {:error, _} ->
+          render(conn, "new_user.html", error: "Something went wrong, please try again")
       end
-    else
-      _ ->
-        changeset =
-          Accounts.change_user(%User{}, user_details)
-          |> Ecto.Changeset.add_error(:captcha, "failed, please try again")
 
-        render(conn, "new_user.html", changeset: changeset, redirect_to: redirect_to)
+    else
+      render(conn, "new_user.html", error: "ReCaptcha failed, please try again")
     end
+
   end
+
 end