+++ lib/oli_web/endpoint.ex
@@ -23,7 +23,7 @@ defmodule OliWeb.Endpoint do
     at: "/",
     from: :oli,
     gzip: false,
-    only: ~w(css fonts images js favicon.ico robots.txt)
+    only: ~w(css fonts images js custom favicons favicon.ico robots.txt)
 
   # Code reloading can be explicitly enabled under the
   # :code_reloader configuration of your endpoint.